ci-tagged

Conflicts:
	bosh-dev/lib/bosh/dev/build.rb
	bosh-dev/spec/unit/bosh/dev/build_spec.rb
	bosh-director/lib/bosh/director/deployment_plan/instance.rb
	bosh-director/lib/bosh/director/jobs/vm_state.rb
	bosh-director/spec/unit/dns_helper_spec.rb
	bosh-director/spec/unit/instance_updater_spec.rb
	bosh-director/spec/unit/models/director_attribute_spec.rb
	bosh-stemcell/Vagrantfile
	bosh-stemcell/lib/bosh/stemcell/stage_collection.rb
	bosh-stemcell/spec/os_image/centos_7_spec.rb
	bosh-stemcell/spec/os_image/ubuntu_trusty_spec.rb
	bosh-stemcell/spec/stemcells/ubuntu_lucid_spec.rb
	bosh-stemcell/spec/stemcells/ubuntu_trusty_spec.rb
	bosh_agent/lib/bosh_agent.rb
	bosh_agent/lib/bosh_agent/apply_plan/job.rb
	bosh_agent/lib/bosh_agent/bootstrap.rb
	bosh_agent/lib/bosh_agent/disk_util.rb
	bosh_agent/lib/bosh_agent/handler.rb
	bosh_agent/lib/bosh_agent/message/apply.rb
	bosh_agent/lib/bosh_agent/message/list_disk.rb
	bosh_agent/lib/bosh_agent/message/migrate_disk.rb
	bosh_agent/lib/bosh_agent/message/mount_disk.rb
	bosh_agent/lib/bosh_agent/message/state.rb
	bosh_agent/lib/bosh_agent/message/unmount_disk.rb
	bosh_agent/lib/bosh_agent/monit.rb
	bosh_agent/lib/bosh_agent/mounter.rb
	bosh_agent/lib/bosh_agent/runner.rb
	bosh_agent/lib/bosh_agent/settings.rb
	bosh_agent/lib/bosh_agent/util.rb
	bosh_agent/spec/unit/bootstrap_spec.rb
	bosh_agent/spec/unit/handler_spec.rb
	bosh_agent/spec/unit/heartbeat_spec.rb
	bosh_agent/spec/unit/message/migrate_disk_spec.rb
	bosh_agent/spec/unit/message/mount_disk_spec.rb
	bosh_agent/spec/unit/message/state_spec.rb
	bosh_agent/spec/unit/message/unmount_disk_spec.rb
	bosh_agent/spec/unit/monit_spec.rb
	bosh_agent/spec/unit/mounter_spec.rb
	bosh_agent/spec/unit/runner_spec.rb
	bosh_vsphere_cpi/lib/cloud/vsphere/cloud.rb
	bosh_vsphere_cpi/lib/cloud/vsphere/vm_creator.rb
	bosh_vsphere_cpi/lib/ruby_vim_sdk/version.rb
	release/.final_builds/jobs/blobstore/index.yml
	release/.final_builds/packages/nginx/index.yml
	release/jobs/director/templates/director.yml.erb.erb
	release/jobs/director/templates/drain
	release/jobs/postgres/spec
	release/spec/director.yml.erb.erb_spec.rb
	stemcell_builder/stages/base_centos_packages/apply.sh
	stemcell_builder/stages/base_ubuntu_firstboot/assets/root/firstboot.sh
	stemcell_builder/stages/base_ubuntu_packages/apply.sh

+++ bosh-director/lib/bosh/director/api/task_remover.rb
@@ -1,13 +1,12 @@
 module Bosh::Director::Api
   class TaskRemover
-    def initialize(max_tasks, logger)
+    def initialize(max_tasks)
       @max_tasks = max_tasks
-      #@logger = logger
     end
 
     def remove
       removal_candidates_dataset.each do |task|
-        FileUtils.rm_rf(task.output)
+        FileUtils.rm_rf(task.output) if task.output
         task.destroy
       end
     end
