+++ bosh-director/lib/bosh/director/jobs/delete_release.rb
@@ -63,7 +63,7 @@ module Bosh::Director
           package.remove_release_version(release_version)
         end
 
-        event_log.begin_stage("Deleting templates", templates_to_delete.count)
+        event_log.begin_stage("Deleting jobs", templates_to_delete.count)
         templates_to_delete.each do |template|
           track_and_log("#{template.name}/#{template.version}") do
             logger.info("Template #{template.name}/#{template.version} " +
@@ -74,7 +74,7 @@ module Bosh::Director
         end
 
         templates_to_keep.each do |template|
-          logger.info("Keeping template " +
+          logger.info("Keeping job " +
                       "#{template.name}/#{template.version} as it is used " +
                       "by other release versions")
           template.remove_release_version(release_version)
@@ -100,7 +100,7 @@ module Bosh::Director
           end
         end
 
-        event_log.begin_stage("Deleting templates", release.templates.count)
+        event_log.begin_stage("Deleting jobs", release.templates.count)
         release.templates.each do |template|
           track_and_log("#{template.name}/#{template.version}") do
             delete_template(template)
@@ -132,21 +132,22 @@ module Bosh::Director
                       "(#{compiled_package.blobstore_id}) " +
                       "#{package.name}/#{package.version} " +
                       "for #{stemcell.name}/#{stemcell.version}")
-          delete_blobstore_id(compiled_package.blobstore_id) do
+
+          if delete_blobstore_id(compiled_package.blobstore_id)
             compiled_package.destroy
           end
         end
 
-        delete_blobstore_id(package.blobstore_id) do
+        if delete_blobstore_id(package.blobstore_id, true)
           package.remove_all_release_versions
           package.destroy
         end
       end
 
       def delete_template(template)
-        logger.info("Deleting template: #{template.name}/#{template.version}")
+        logger.info("Deleting job: #{template.name}/#{template.version}")
 
-        delete_blobstore_id(template.blobstore_id) do
+        if delete_blobstore_id(template.blobstore_id)
           template.remove_all_release_versions
           template.destroy
         end
@@ -201,17 +202,21 @@ module Bosh::Director
         "/release/#{@name}"
       end
 
-      def delete_blobstore_id(blobstore_id)
+      def delete_blobstore_id(blobstore_id, nil_id_allowed = false)
+        if blobstore_id.nil? && nil_id_allowed
+          return true
+        end
+
         deleted = false
         begin
           @blobstore.delete(blobstore_id)
           deleted = true
         rescue Exception => e
-          logger.warn("Could not delete from blobstore: #{e}\n " +
-                      e.backtrace.join("\n"))
+          logger.warn("Could not delete from blobstore: #{e}\n " + e.backtrace.join("\n"))
           @errors << e
         end
-        yield if deleted || @force
+
+        return deleted || @force
       end
 
     end
